Start migration to classes instead of module + Improve coverage report
