+++ .github/workflows/python-test.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: ['3.6', '3.7', '3.8', '3.9', '3.10']
         mongodb-version: ['4.4']
 
     steps:
@@ -33,6 +33,7 @@ jobs:
         mongodb-version: ${{ matrix.mongodb-version }}
     - name: Ensure right config
       run: |
+        mv "./coverage.xml" "./sql_coverage.xml"
         mv "./flask_monitoringdashboard/core/config/__init__.py" "./flask_monitoringdashboard/core/config/__init__.py.old"
         mv "./flask_monitoringdashboard/core/config/__init__.py.mongodb" "./flask_monitoringdashboard/core/config/__init__.py"
     - name: Test with pytest with mongodb
@@ -41,4 +42,4 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
       with:
-        file: ./coverage.xml
+        file: ./*coverage.xml
