+++ flask_monitoringdashboard/controllers/versions.py
@@ -1,11 +1,10 @@
 import numpy
 
-from flask_monitoringdashboard.database import Request
 from flask_monitoringdashboard.database.count_group import get_value, count_requests_group
 from flask_monitoringdashboard.database.data_grouped import get_two_columns_grouped
 from flask_monitoringdashboard.database.endpoint import get_endpoint_by_name
 from flask_monitoringdashboard.database.versions import get_first_requests, \
-    get_2d_version_data_filter
+    get_2d_version_data_filter, get_field_name, get_version_requested_query
 
 
 def get_2d_version_data(session, endpoint_id, versions, column_data, column):
@@ -28,13 +27,11 @@ def get_2d_version_data(session, endpoint_id, versions, column_data, column):
 
 
 def get_version_user_data(session, endpoint_id, versions, users):
-    return get_2d_version_data(session, endpoint_id, versions, users, Request.group_by
-                               if not getattr(Request, "is_mongo_db", False) else "group_by")
+    return get_2d_version_data(session, endpoint_id, versions, users, get_field_name("group_by"))
 
 
 def get_version_ip_data(session, endpoint_id, versions, ips):
-    return get_2d_version_data(session, endpoint_id, versions, ips, Request.ip
-                               if not getattr(Request, "is_mongo_db", False) else "ip")
+    return get_2d_version_data(session, endpoint_id, versions, ips, get_field_name("ip"))
 
 
 def get_multi_version_data(session, endpoints, versions):
@@ -47,8 +44,7 @@ def get_multi_version_data(session, endpoints, versions):
     """
     endpoints = [get_endpoint_by_name(session, name) for name in endpoints]
     requests = [count_requests_group(session,
-                                     Request.version_requested == v if not getattr(Request, "is_mongo_db", False) else
-                                     {"version_requested": v})
+                                     get_version_requested_query(v))
                 for v in versions]
 
     total_hits = numpy.zeros(len(versions))
