+++ flask_monitoringdashboard/core/custom_graph/__init__.py
@@ -3,7 +3,7 @@ import os
 
 from apscheduler.schedulers import SchedulerAlreadyRunningError
 from apscheduler.schedulers.background import BackgroundScheduler
-
+from flask_monitoringdashboard.core.logger import log
 from flask_monitoringdashboard.database import session_scope
 from flask_monitoringdashboard.database.custom_graph import (
     add_value,
@@ -18,10 +18,10 @@ def init(app):
     if not app.debug or os.environ.get('WERKZEUG_RUN_MAIN') == 'true':
         try:
             scheduler.start()
-            print('Scheduler started')
+            log('Scheduler started')
             atexit.register(lambda: scheduler.shutdown())
         except SchedulerAlreadyRunningError as err:
-            print(err)
+            log(err)
 
 
 def register_graph(name):
