+++ flask_monitoringdashboard/database/auth.py
@@ -1,44 +1,29 @@
-from flask_monitoringdashboard.database import User, session_scope
+from flask_monitoringdashboard.database import User, session_scope, UserQueries
 from flask_monitoringdashboard import config
-from sqlalchemy.orm.exc import NoResultFound
-from sqlalchemy.exc import IntegrityError
-from pymongo.errors import DuplicateKeyError
 
 
 def update_user(new_password, old_password, user_id, user_is_admin):
     with session_scope() as session:
-        if not getattr(User, "is_mongo_db", False):
-            user = session.query(User).filter(User.id == user_id).one()
-        else:
-            try:
-                user = User(**User().get_collection(session).find_one({"id": user_id}))
-            except TypeError:
-                raise NoResultFound()
+        data_base_operation = UserQueries(session)
+        user = data_base_operation.find_by_id(User, user_id)
         user.is_admin = user_is_admin
         if old_password:
             if user.check_password(old_password):
                 user.set_password(new_password)
-                if getattr(User, "is_mongo_db", False):
-                    user.get_collection(session).update_one({"id": user_id}, {"$set": user})
+                data_base_operation.finalize_update(user)
                 return True
             else:
                 return False
-        if getattr(User, "is_mongo_db", False):
-            user.get_collection(session).update_one({"id": user_id}, {"$set": user})
+        data_base_operation.finalize_update(user)
         return True
 
 
 def create_default_user(session):
-    if getattr(User, "is_mongo_db", False):
-        if User().get_collection(session).count_documents({}) == 0:
-            user = User(username=config.username, is_admin=True)
-            user.set_password(password=config.password)
-            user.get_collection(session).insert_one(user)
-    else:
-        if session.query(User).count() == 0:
-            user = User(username=config.username, is_admin=True)
-            user.set_password(password=config.password)
-            session.add(user)
+    data_base_operation = UserQueries(session)
+    if data_base_operation.count(User) == 0:
+        user = User(username=config.username, is_admin=True)
+        user.set_password(password=config.password)
+        data_base_operation.create_obj(user)
 
 
 def get_user(username, password):
@@ -47,19 +32,12 @@ def get_user(username, password):
     """
     with session_scope() as session:
         create_default_user(session)
-        if getattr(User, "is_mongo_db", False):
-            try:
-                user = User(**User().get_collection(session).find_one({"username": username}))
-            except TypeError:
-                user = None
-        else:
-            user = session.query(User).filter(User.username == username).one_or_none()
+        database_operation = UserQueries(session)
+        user = database_operation.find_one_user_or_none(username=username)
         if user is not None:
             if user.check_password(password=password):
-                if not getattr(User, "is_mongo_db", False):
-                    session.expunge_all()
+                database_operation.expunge_all()
                 return user
-
     return None
 
 
@@ -67,31 +45,19 @@ def add_user(username, is_admin, password):
     with session_scope() as session:
         user = User(username=username, is_admin=is_admin)
         user.set_password(password=password)
-        if getattr(User, "is_mongo_db", False):
-            try:
-                user.get_collection(session).insert_one(user)
-            except DuplicateKeyError:
-                raise IntegrityError(orig="add_user", params="user_id", statement="USER ALREADY EXIST")
-        else:
-            session.add(user)
-            session.commit()
+        data_base_operation = UserQueries(session)
+        data_base_operation.create_obj(user)
+        data_base_operation.commit()
 
 
 def delete_user(user_id):
     with session_scope() as session:
-        if getattr(User, "is_mongo_db", False):
-            User().get_collection(session).delete_one({"id": user_id})
-        else:
-            session.query(User).filter(User.id == user_id).delete()
+        UserQueries(session).delete_user(user_id)
 
 
 def get_all_users():
     with session_scope() as session:
-        if getattr(User, "is_mongo_db", False):
-            users = list(User(**elem) for elem in User().get_collection(session).find({}).sort([("id", -1)]))
-        else:
-            users = session.query(User).order_by(User.id).all()
-
+        users = UserQueries(session).find_all_user()
         return [
             {
                 'id': user.id,
