+++ flask_monitoringdashboard/database/count.py
@@ -1,6 +1,4 @@
-from sqlalchemy import func, distinct
-
-from flask_monitoringdashboard.database import Request, StackLine, Outlier
+from flask_monitoringdashboard.database import CountQueries
 
 
 def count_rows(session, column, *criterion):
@@ -11,12 +9,7 @@ def count_rows(session, column, *criterion):
     :param criterion: where-clause of the query
     :return: number of rows
     """
-
-    if getattr(Request, "is_mongo_db", False):
-        return Request().get_collection(session).count_documents({"$and": list(criterion)}
-                                                                 if len(criterion) > 0 else {})
-    else:
-        return session.query(func.count(distinct(column))).filter(*criterion).scalar()
+    return CountQueries(session).count_rows(column, *criterion)
 
 
 def count_requests(session, endpoint_id, *where):
@@ -26,11 +19,7 @@ def count_requests(session, endpoint_id, *where):
     :param endpoint_id: id of the endpoint
     :param where: additional arguments
     """
-    return count_rows(session,
-                      Request.id if not getattr(Request, "is_mongo_db", False) else "id",
-                      Request.endpoint_id == endpoint_id if not getattr(Request, "is_mongo_db", False) else
-                      {"endpoint_id": endpoint_id},
-                      *where)
+    return CountQueries(session).count_requests(endpoint_id, *where)
 
 
 def count_total_requests(session, *where):
@@ -39,9 +28,7 @@ def count_total_requests(session, *where):
     :param session: session for the database
     :param where: additional arguments
     """
-    return count_rows(session,
-                      Request.id if not getattr(Request, "is_mongo_db", False) else "id",
-                      *where)
+    return CountQueries(session).count_total_requests(*where)
 
 
 def count_outliers(session, endpoint_id):
@@ -50,13 +37,7 @@ def count_outliers(session, endpoint_id):
     :param endpoint_id: id of the endpoint
     :return: An integer with the number of rows in the Outlier-table.
     """
-    if getattr(Request, "is_mongo_db", False):
-        return Outlier().get_collection(session).count_documents({"endpoint_id": endpoint_id})
-    else:
-        return count_rows(session,
-                          Request.id,
-                          Request.endpoint_id == endpoint_id,
-                          Request.outlier)
+    return CountQueries(session).count_outliers(endpoint_id)
 
 
 def count_profiled_requests(session, endpoint_id):
@@ -66,12 +47,4 @@ def count_profiled_requests(session, endpoint_id):
     :param endpoint_id: id of the endpoint
     :return: An integer
     """
-    if getattr(Request, "is_mongo_db", False):
-        return len(StackLine().get_collection(session).distinct("request_id", {"endpoint_id": endpoint_id}))
-    else:
-        return (
-            session.query(func.count(distinct(StackLine.request_id)))
-            .filter(Request.endpoint_id == endpoint_id)
-            .join(Request.stack_lines)
-            .scalar()
-        )
+    return CountQueries(session).count_profiled_requests(endpoint_id)
