+++ flask_monitoringdashboard/database/data_grouped.py
@@ -1,6 +1,6 @@
 from numpy import median
 
-from flask_monitoringdashboard.database import Request
+from flask_monitoringdashboard.database import Request, CountQueries
 
 
 def get_data_grouped(session, column, func, *where):
@@ -10,15 +10,7 @@ def get_data_grouped(session, column, func, *where):
     :param func: the function to reduce the data
     :param where: additional where clause
     """
-
-    if getattr(Request, "is_mongo_db", False):
-        result = list((elem[column], elem["duration"]) for elem in
-                      Request().get_collection(session).find({"$and": list(where)}
-                                                             if len(where) > 0 else {}).sort([(column, 1)]))
-    else:
-        result = session.query(column, Request.duration).filter(*where).order_by(column).all()
-        # result is now a list of tuples per request.
-    return group_result(result, func)
+    return group_result(CountQueries(session).get_data_grouped(column, *where), func)
 
 
 def group_result(result, func):
@@ -62,7 +54,7 @@ def get_endpoint_data_grouped(session, func, *where):
     :param where: additional where clause
     """
     return get_data_grouped(session,
-                            Request.endpoint_id if not getattr(Request, "is_mongo_db", False) else "endpoint_id",
+                            CountQueries.get_field_name("endpoint_id", Request),
                             func,
                             *where)
 
@@ -74,8 +66,7 @@ def get_version_data_grouped(session, func, *where):
     :param where: additional where clause
     """
     return get_data_grouped(session,
-                            Request.version_requested if not getattr(Request, "is_mongo_db", False) else
-                            "version_requested",
+                            CountQueries.get_field_name("version_requested", Request),
                             func,
                             *where)
 
@@ -87,7 +78,7 @@ def get_user_data_grouped(session, func, *where):
     :param where: additional where clause
     """
     return get_data_grouped(session,
-                            Request.group_by if not getattr(Request, "is_mongo_db", False) else "group_by",
+                            CountQueries.get_field_name("group_by", Request),
                             func,
                             *where)
 
@@ -98,12 +89,4 @@ def get_two_columns_grouped(session, column, *where):
     :param column: column that is used for the grouping (together with the Request.version)
     :param where: additional where clause
     """
-    if getattr(Request, "is_mongo_db", False):
-        result = list(((elem[column], elem["version_requested"]), elem["duration"]) for elem in
-                      Request().get_collection(session).find({"$and": list(where)}).sort([(column, 1)]))
-    else:
-        result = (
-            session.query(column, Request.version_requested, Request.duration).filter(*where).all()
-        )
-        result = [((g, v), t) for g, v, t in result]
-    return group_result(result, median)
+    return group_result(CountQueries(session).get_two_columns_grouped(column, *where), median)
