+++ flask_monitoringdashboard/database/versions.py
@@ -1,6 +1,12 @@
-from sqlalchemy import func, desc
+from flask_monitoringdashboard.database import Request, VersionQuery
 
-from flask_monitoringdashboard.database import Request
+
+def get_field_name(name):
+    return VersionQuery.get_field_name(name, Request)
+
+
+def get_version_requested_query(v):
+    return VersionQuery.get_version_requested_query(v)
 
 
 def get_versions(session, endpoint_id=None, limit=None):
@@ -11,35 +17,11 @@ def get_versions(session, endpoint_id=None, limit=None):
     :param limit: only return the most recent versions
     :return: a list of tuples with the versions (as a string) and dates, from oldest to newest
     """
-    if getattr(Request, "is_mongo_db", False):
-        query = [
-            {"$group": {
-                "_id": "$version_requested",
-                "minTime": {"$min": "$time_requested"}
-            }},
-            {"$sort": {"minTime": -1}}
-        ]
-        if endpoint_id:
-            query.insert(0, {"$match": {"endpoint_id": endpoint_id}})
-        if limit:
-            query.append({"$limit": int(limit)})
-        return list((str(elem["_id"]), elem["minTime"]) for elem in Request().get_collection(session).aggregate(query))
-    else:
-        query = session.query(Request.version_requested, func.min(Request.time_requested))
-        if endpoint_id:
-            query = query.filter(Request.endpoint_id == endpoint_id)
-        query = query.group_by(Request.version_requested)
-        query = query.order_by(func.min(Request.time_requested).desc())
-        if limit:
-            query = query.limit(limit)
-        return query.all()
+    return VersionQuery(session).get_versions(endpoint_id=endpoint_id, limit=limit)
 
 
 def get_2d_version_data_filter(endpoint_id):
-    if getattr(Request, "is_mongo_db", False):
-        return {"endpoint_id": endpoint_id}
-    else:
-        return Request.endpoint_id == endpoint_id
+    return VersionQuery.get_2d_version_data_filter(endpoint_id)
 
 
 def get_first_requests(session, endpoint_id, limit=None):
@@ -50,27 +32,4 @@ def get_first_requests(session, endpoint_id, limit=None):
     :param endpoint_id: id of the endpoint
     :return list of tuples with versions
     """
-    if getattr(Request, "is_mongo_db", False):
-        query = [
-            {"$match": {"endpoint_id": endpoint_id}},
-            {"$group": {
-                "_id": "$version_requested",
-                "minTime": {"$min": "$time_requested"}
-            }},
-            {"$sort": {"minTime": -1}}
-        ]
-        if limit:
-            query.append({"$limit": int(limit)})
-        return list((elem["_id"], elem["minTime"]) for elem in Request().get_collection(session).aggregate(query))
-    else:
-        query = (
-            session.query(
-                Request.version_requested, func.min(Request.time_requested).label('first_used')
-            )
-                .filter(Request.endpoint_id == endpoint_id)
-                .group_by(Request.version_requested)
-                .order_by(desc('first_used'))
-        )
-        if limit:
-            query = query.limit(limit)
-        return query.all()
+    return VersionQuery(session).get_first_requests(endpoint_id, limit=limit)
